Django-Rest-Framework/requirements.txt
django==4.1.0
djangorestframework
pyyaml
requests
django-cors-headers